+++ controllers/ServiceAdminController.php
@@ -8,7 +8,9 @@
 namespace anli\auth0\controllers;
 
 use anli\auth0\models\ApiUser;
+use anli\auth0\models\Tenant;
 use Yii;
+use yii\filters\AccessControl;
 
 /**
  * This is the ServiceAdmin controller class for Auth0.
@@ -18,6 +20,30 @@ use Yii;
 class ServiceAdminController extends \yii\web\Controller
 {
     /**
+     * @inheritdoc
+     */
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'rules' => [
+                    [
+                        'actions' => [
+                            'index',
+                        ],
+                        'allow' => true,
+                        'roles' => ['@'],
+                        'matchCallback' => function ($rule, $action) {
+                            return Yii::$app->controller->module->isAdmin;
+                        }
+                    ],
+                ],
+            ],
+        ];
+    }
+
+    /**
      * See all users by service
      * @return mixed
      */
@@ -25,10 +51,12 @@ class ServiceAdminController extends \yii\web\Controller
     {
         Yii::$app->user->setReturnUrl(['/' . $this->getRoute()]);
 
-        $query = ApiUser::find()->orderBy('email:1');
+        $userQuery = ApiUser::find()->orderBy('email:1');
+        $tenantQuery = Tenant::find()->orderBy('name');
 
         return $this->render('index', [
-            'query' => $query,
+            'userQuery' => $userQuery,
+            'tenantQuery' => $tenantQuery,
         ]);
     }
 }