+++ models/ApiUserColumn.php
@@ -54,17 +54,18 @@ class ApiUserColumn
     }
 
     /**
+     * @param string $tenantName
      * @return mixed
      */
-    public function role()
+    public function role($tenantName)
     {
         $this->columns = array_merge($this->columns, [
             [
                 'label' => 'Role',
-                'value' => function ($model, $key, $index, $column) {
-                    if (isset($model['app_metadata']['permissions'][Yii::$app->getModule('auth0')->serviceId][Yii::$app->tenant->identity->name]['role'])) {
+                'value' => function ($model, $key, $index, $column) use ($tenantName) {
+                    if (isset($model['app_metadata']['permissions'][Yii::$app->getModule('auth0')->serviceId][$tenantName]['role'])) {
 
-                        return $model['app_metadata']['permissions'][Yii::$app->getModule('auth0')->serviceId][Yii::$app->tenant->identity->name]['role'];
+                        return $model['app_metadata']['permissions'][Yii::$app->getModule('auth0')->serviceId][$tenantName]['role'];
                     }
                     return '';
                 }
@@ -83,33 +84,69 @@ class ApiUserColumn
     }
 
     /**
-     * @return string $template
+     * @param string $template
+     * @param mixed $params
      * @return mixed
      */
-    public function actions($template = '{update} {delete}')
+    public function actions($template = '{update} {delete}', $params = '')
     {
         $this->columns = array_merge($this->columns, [
             [
-                'header' => 'actions',
                 'class' => ActionColumn::className(),
                 'controller' => SELF::CONTROLLER,
                 'template' => $template,
                 'buttons' => [
-                    'update-role-to-user' => function ($url, $model, $key) {
-                        return Html::a('add', [SELF::CONTROLLER . '/update-role', 'userId' => $model['user_id'], 'role' => 'user'], [
-                            'title' => 'Add',
+                    'update-role-to-user' => function ($url, $model, $key) use ($params) {
+                        return Html::a('<i class="fa fa-check"></i>', [SELF::CONTROLLER . '/update-role', 'userId' => $model['user_id'], 'tenantId' => $params, 'role' => 'user'], [
+                            'title' => 'Update to User',
+                            'data-toggle' => 'tooltip',
+                        ]);
+                    },
+                    'update-role-to-admin' => function ($url, $model, $key) use ($params) {
+                        return Html::a('<i class="fa fa-check-circle"></i>', [SELF::CONTROLLER . '/update-role', 'userId' => $model['user_id'], 'tenantId' => $params, 'role' => 'admin'], [
+                            'title' => 'Update to Admin',
                             'data-toggle' => 'tooltip',
                         ]);
                     },
                     'remove-tenant' => function ($url, $model, $key) {
-                        return Html::a('remove', [SELF::CONTROLLER . '/remove-tenant', 'userId' => $model['user_id']], [
+                        return Html::a('<i class="fa fa-times"></i>', [SELF::CONTROLLER . '/remove-tenant', 'userId' => $model['user_id']], [
                             'title' => 'Remove',
                             'data-toggle' => 'tooltip',
                         ]);
                     },
                 ],
+                'contentOptions' => ['class' => 'text-right'],
             ]
         ]);
         return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function tenants()
+    {
+        $this->columns = array_merge($this->columns, [
+            [
+                'label' => 'Tenants',
+                'value' => function ($model, $key, $index, $column) {
+
+                    $count = TenantUser::find()->joinWith(['user.auth'])->andWhere(['{{%auth}}.source_id' => $model['user_id']])->count();
+
+                    return ($count > 0) ? Yii::$app->formatter->asDecimal($count, 0) : '' ;
+                },
+                'contentOptions' => ['class' => 'text-center'],
+                'headerOptions' => ['class' => 'text-center'],
+            ],
+        ]);
+        return $this;
+    }
+
+    /**
+     * @return \yii\db\ActiveQuery
+     */
+    public function getTenantUsers()
+    {
+        return $this->hasMany(TenantUser::className(), ['tenant_id' => 'id']);
     }
 }