+++ views/tenant-admin/index.php
@@ -1,8 +1,0 @@
-<?php
-
-use yii\helpers\Html;
-
-/* @var $this yii\web\View */
-?>
-
-<?= \anli\auth0\widgets\ApiUserGridView::widget(['query' => $query]);?>