+++ CHANGELOG.md
@@ -1,6 +1,33 @@
 CHANGELOG
 =====
 
+1.3.0 [28 Aug 2015]
+-----
+
+* `enh` Added user select2Data function
+* `enh` Added tenant function
+* `chg` Changed login action with `goBack()` function
+* `enh` Added access control to controller
+* `enh` Added documentation on admin email configuration
+* `enh` Added update user feature to admin
+
+1.2.0 [19 Aug 2015]
+-----
+
+* `enh` Added metronic theme
+* `enh` Added create action and create form with modal assets
+* `enh` Added buttons to tenant
+* `enh` Added login layout
+* `enh` Added confirmation msg for delete all, added audit log for tenant
+* `enh` Added users column to tenant portlet
+* `enh` Added user portlet in tenant admin dashboard with tenants column
+* `chg` Removed tick and cross in service-admin index
+* `chg` Changed portlet layout similar to the old design
+* `enh` Added the import all function in tenant service admin
+* `enh` Added export function to tenant service admin
+* `enh` Added session flash feature
+* `enh` Added checkbox function to gridview in service admin
+
 1.1.0 [12 Aug 2015]
 -----